perf: Separate swap constants to reduce bundle
