+++ src/__tests__/utils/index.test.ts
@@ -1,7 +1,8 @@
 import { BigNumber } from '@ethersproject/bignumber'
 import { AddressZero } from '@ethersproject/constants'
 import { TokenAmount, Token, ChainId, Percent, JSBI } from '@pancakeswap/sdk'
-import { getBscScanLink, calculateSlippageAmount, isAddress, calculateGasMargin, basisPointsToPercent } from 'utils'
+import { getBscScanLink, isAddress, calculateGasMargin } from 'utils'
+import { calculateSlippageAmount, basisPointsToPercent } from 'utils/exchange'
 
 describe('utils', () => {
   describe('#getBscScanLink', () => {
