+++ src/__tests__/utils/prices.test.ts
@@ -1,5 +1,5 @@
 import { ChainId, JSBI, Pair, Route, Token, TokenAmount, Trade, TradeType } from '@pancakeswap/sdk'
-import { computeTradePriceBreakdown } from 'utils/prices'
+import { computeTradePriceBreakdown } from 'utils/exchange'
 
 describe('prices', () => {
   const token1 = new Token(ChainId.MAINNET, '0x0000000000000000000000000000000000000001', 18)
