+++ src/hooks/limitOrders/useGasOverhead.ts
@@ -6,7 +6,7 @@ import { useTradeExactIn } from 'hooks/Trades'
 import tryParseAmount from 'utils/tryParseAmount'
 import { Rate } from 'state/limitOrders/types'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
-import { GENERIC_GAS_LIMIT_ORDER_EXECUTION, BIG_INT_TEN } from 'config/constants'
+import { GENERIC_GAS_LIMIT_ORDER_EXECUTION, BIG_INT_TEN } from 'config/constants/exchange'
 import getGasPrice from 'utils/getGasPrice'
 import getPriceForOneToken from 'views/LimitOrders/utils/getPriceForOneToken'
