+++ src/hooks/useApproveCallback.ts
@@ -5,11 +5,11 @@ import { CHAIN_ID } from 'config/constants/networks'
 import { useCallback, useMemo } from 'react'
 import { logError } from 'utils/sentry'
 import { useWeb3React } from '@web3-react/core'
-import { ROUTER_ADDRESS } from '../config/constants'
+import { ROUTER_ADDRESS } from 'config/constants/exchange'
 import useTokenAllowance from './useTokenAllowance'
 import { Field } from '../state/swap/actions'
 import { useTransactionAdder, useHasPendingApproval } from '../state/transactions/hooks'
-import { computeSlippageAdjustedAmounts } from '../utils/prices'
+import { computeSlippageAdjustedAmounts } from '../utils/exchange'
 import { calculateGasMargin } from '../utils'
 import { useTokenContract } from './useContract'
 import { useCallWithGasPrice } from './useCallWithGasPrice'
