+++ src/state/swap/hooks.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-var */
 /* eslint-disable vars-on-top */
 import { renderHook } from '@testing-library/react-hooks'
-import { DEFAULT_OUTPUT_CURRENCY } from 'config/constants'
+import { DEFAULT_OUTPUT_CURRENCY } from 'config/constants/exchange'
 import { parse } from 'querystring'
 import { useCurrency } from 'hooks/Tokens'
 import { createReduxWrapper } from 'testUtils'
