+++ src/state/swap/useSwapActionHandlers.ts
@@ -1,0 +1,49 @@
+import { useCallback } from 'react'
+import { Currency, Token, ETHER } from '@pancakeswap/sdk'
+import { Field, selectCurrency, switchCurrencies, typeInput, setRecipient } from './actions'
+import { useAppDispatch } from '../index'
+
+export function useSwapActionHandlers(): {
+  onCurrencySelection: (field: Field, currency: Currency) => void
+  onSwitchTokens: () => void
+  onUserInput: (field: Field, typedValue: string) => void
+  onChangeRecipient: (recipient: string | null) => void
+} {
+  const dispatch = useAppDispatch()
+  const onCurrencySelection = useCallback(
+    (field: Field, currency: Currency) => {
+      dispatch(
+        selectCurrency({
+          field,
+          currencyId: currency instanceof Token ? currency.address : currency === ETHER ? 'BNB' : '',
+        }),
+      )
+    },
+    [dispatch],
+  )
+
+  const onSwitchTokens = useCallback(() => {
+    dispatch(switchCurrencies())
+  }, [dispatch])
+
+  const onUserInput = useCallback(
+    (field: Field, typedValue: string) => {
+      dispatch(typeInput({ field, typedValue }))
+    },
+    [dispatch],
+  )
+
+  const onChangeRecipient = useCallback(
+    (recipient: string | null) => {
+      dispatch(setRecipient({ recipient }))
+    },
+    [dispatch],
+  )
+
+  return {
+    onSwitchTokens,
+    onCurrencySelection,
+    onUserInput,
+    onChangeRecipient,
+  }
+}
