+++ src/utils/exchange.ts
@@ -1,0 +1,109 @@
+import { Web3Provider } from '@ethersproject/providers'
+import IPancakeRouter02ABI from 'config/abi/IPancakeRouter02.json'
+import { IPancakeRouter02 } from 'config/abi/types/IPancakeRouter02'
+import { CHAIN_ID } from 'config/constants/networks'
+import { JSBI, Percent, CurrencyAmount, Trade, Fraction, TokenAmount } from '@pancakeswap/sdk'
+import {
+  ROUTER_ADDRESS,
+  BIPS_BASE,
+  ONE_HUNDRED_PERCENT,
+  INPUT_FRACTION_AFTER_FEE,
+  BLOCKED_PRICE_IMPACT_NON_EXPERT,
+  ALLOWED_PRICE_IMPACT_HIGH,
+  ALLOWED_PRICE_IMPACT_MEDIUM,
+  ALLOWED_PRICE_IMPACT_LOW,
+} from 'config/constants/exchange'
+import { getContract, getProviderOrSigner } from './index'
+import { Field } from '../state/swap/actions'
+
+// converts a basis points value to a sdk percent
+export function basisPointsToPercent(num: number): Percent {
+  return new Percent(JSBI.BigInt(num), BIPS_BASE)
+}
+
+export function calculateSlippageAmount(value: CurrencyAmount, slippage: number): [JSBI, JSBI] {
+  if (slippage < 0 || slippage > 10000) {
+    throw Error(`Unexpected slippage value: ${slippage}`)
+  }
+  return [
+    JSBI.divide(JSBI.multiply(value.raw, JSBI.BigInt(10000 - slippage)), BIPS_BASE),
+    JSBI.divide(JSBI.multiply(value.raw, JSBI.BigInt(10000 + slippage)), BIPS_BASE),
+  ]
+}
+
+// account is optional
+export function getRouterContract(_: number, library: Web3Provider, account?: string) {
+  return getContract(
+    ROUTER_ADDRESS[CHAIN_ID],
+    IPancakeRouter02ABI,
+    getProviderOrSigner(library, account),
+  ) as IPancakeRouter02
+}
+
+// computes price breakdown for the trade
+export function computeTradePriceBreakdown(trade?: Trade | null): {
+  priceImpactWithoutFee: Percent | undefined
+  realizedLPFee: CurrencyAmount | undefined | null
+} {
+  // for each hop in our trade, take away the x*y=k price impact from 0.3% fees
+  // e.g. for 3 tokens/2 hops: 1 - ((1 - .03) * (1-.03))
+  const realizedLPFee = !trade
+    ? undefined
+    : ONE_HUNDRED_PERCENT.subtract(
+        trade.route.pairs.reduce<Fraction>(
+          (currentFee: Fraction): Fraction => currentFee.multiply(INPUT_FRACTION_AFTER_FEE),
+          ONE_HUNDRED_PERCENT,
+        ),
+      )
+
+  // remove lp fees from price impact
+  const priceImpactWithoutFeeFraction = trade && realizedLPFee ? trade.priceImpact.subtract(realizedLPFee) : undefined
+
+  // the x*y=k impact
+  const priceImpactWithoutFeePercent = priceImpactWithoutFeeFraction
+    ? new Percent(priceImpactWithoutFeeFraction?.numerator, priceImpactWithoutFeeFraction?.denominator)
+    : undefined
+
+  // the amount of the input that accrues to LPs
+  const realizedLPFeeAmount =
+    realizedLPFee &&
+    trade &&
+    (trade.inputAmount instanceof TokenAmount
+      ? new TokenAmount(trade.inputAmount.token, realizedLPFee.multiply(trade.inputAmount.raw).quotient)
+      : CurrencyAmount.ether(realizedLPFee.multiply(trade.inputAmount.raw).quotient))
+
+  return { priceImpactWithoutFee: priceImpactWithoutFeePercent, realizedLPFee: realizedLPFeeAmount }
+}
+
+// computes the minimum amount out and maximum amount in for a trade given a user specified allowed slippage in bips
+export function computeSlippageAdjustedAmounts(
+  trade: Trade | undefined,
+  allowedSlippage: number,
+): { [field in Field]?: CurrencyAmount } {
+  const pct = basisPointsToPercent(allowedSlippage)
+  return {
+    [Field.INPUT]: trade?.maximumAmountIn(pct),
+    [Field.OUTPUT]: trade?.minimumAmountOut(pct),
+  }
+}
+
+export function warningSeverity(priceImpact: Percent | undefined): 0 | 1 | 2 | 3 | 4 {
+  if (!priceImpact?.lessThan(BLOCKED_PRICE_IMPACT_NON_EXPERT)) return 4
+  if (!priceImpact?.lessThan(ALLOWED_PRICE_IMPACT_HIGH)) return 3
+  if (!priceImpact?.lessThan(ALLOWED_PRICE_IMPACT_MEDIUM)) return 2
+  if (!priceImpact?.lessThan(ALLOWED_PRICE_IMPACT_LOW)) return 1
+  return 0
+}
+
+export function formatExecutionPrice(trade?: Trade, inverted?: boolean): string {
+  if (!trade) {
+    return ''
+  }
+  return inverted
+    ? `${trade.executionPrice.invert().toSignificant(6)} ${trade.inputAmount.currency.symbol} / ${
+        trade.outputAmount.currency.symbol
+      }`
+    : `${trade.executionPrice.toSignificant(6)} ${trade.outputAmount.currency.symbol} / ${
+        trade.inputAmount.currency.symbol
+      }`
+}
