+++ src/utils/index.ts
@@ -5,12 +5,9 @@ import { getAddress } from '@ethersproject/address'
 import { AddressZero } from '@ethersproject/constants'
 import { JsonRpcSigner, Web3Provider } from '@ethersproject/providers'
 import { BigNumber } from '@ethersproject/bignumber'
-import IPancakeRouter02ABI from 'config/abi/IPancakeRouter02.json'
-import { IPancakeRouter02 } from 'config/abi/types/IPancakeRouter02'
 import { CHAIN_ID } from 'config/constants/networks'
-import { JSBI, Percent, Token, CurrencyAmount, Currency, ETHER } from '@pancakeswap/sdk'
+import { Token, Currency, ETHER } from '@pancakeswap/sdk'
 import { TokenAddressMap } from 'state/types'
-import { ROUTER_ADDRESS, BIPS_BASE } from '../config/constants'
 import { BASE_BSC_SCAN_URLS } from '../config'
 import { simpleRpcProvider } from './providers'
 
@@ -58,21 +55,6 @@ export function calculateGasMargin(value: BigNumber): BigNumber {
   return value.mul(BigNumber.from(10000).add(BigNumber.from(1000))).div(BigNumber.from(10000))
 }
 
-// converts a basis points value to a sdk percent
-export function basisPointsToPercent(num: number): Percent {
-  return new Percent(JSBI.BigInt(num), BIPS_BASE)
-}
-
-export function calculateSlippageAmount(value: CurrencyAmount, slippage: number): [JSBI, JSBI] {
-  if (slippage < 0 || slippage > 10000) {
-    throw Error(`Unexpected slippage value: ${slippage}`)
-  }
-  return [
-    JSBI.divide(JSBI.multiply(value.raw, JSBI.BigInt(10000 - slippage)), BIPS_BASE),
-    JSBI.divide(JSBI.multiply(value.raw, JSBI.BigInt(10000 + slippage)), BIPS_BASE),
-  ]
-}
-
 // account is not optional
 export function getSigner(library: Web3Provider, account: string): JsonRpcSigner {
   return library.getSigner(account).connectUnchecked()
@@ -90,15 +72,6 @@ export function getContract(address: string, ABI: any, signer?: Signer | Provide
   }
 
   return new Contract(address, ABI, signer ?? simpleRpcProvider)
-}
-
-// account is optional
-export function getRouterContract(_: number, library: Web3Provider, account?: string) {
-  return getContract(
-    ROUTER_ADDRESS[CHAIN_ID],
-    IPancakeRouter02ABI,
-    getProviderOrSigner(library, account),
-  ) as IPancakeRouter02
 }
 
 export function escapeRegExp(string: string): string {
