+++ src/views/AddLiquidity/PoolPriceBar.tsx
@@ -1,9 +1,9 @@
 import { Currency, Percent, Price } from '@pancakeswap/sdk'
 import { Text } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
+import { ONE_BIPS } from 'config/constants/exchange'
 import { AutoColumn } from '../../components/Layout/Column'
 import { AutoRow } from '../../components/Layout/Row'
-import { ONE_BIPS } from '../../config/constants'
 import { Field } from '../../state/mint/actions'
 
 function PoolPriceBar({
