+++ src/views/LimitOrders/index.tsx
@@ -16,7 +16,8 @@ import { ApprovalState, useApproveCallbackFromInputCurrencyAmount } from 'hooks/
 import { Field } from 'state/limitOrders/types'
 import { useDefaultsFromURLSearch } from 'state/limitOrders/hooks'
 import { maxAmountSpend } from 'utils/maxAmountSpend'
-import { GELATO_NATIVE, LIMIT_ORDERS_DOCS_URL } from 'config/constants'
+import { GELATO_NATIVE } from 'config/constants'
+import { LIMIT_ORDERS_DOCS_URL } from 'config/constants/exchange'
 import { useExchangeChartManager } from 'state/user/hooks'
 import PriceChartContainer from 'views/Swap/components/Chart/PriceChartContainer'
 import { useWeb3React } from '@web3-react/core'
