+++ src/views/Swap/components/FormattedPriceImpact.tsx
@@ -1,6 +1,6 @@
 import { Percent } from '@pancakeswap/sdk'
-import { warningSeverity } from 'utils/prices'
-import { ONE_BIPS } from '../../../config/constants'
+import { warningSeverity } from 'utils/exchange'
+import { ONE_BIPS } from 'config/constants/exchange'
 import { ErrorText } from './styleds'
 
 /**
