+++ src/views/Swap/index.tsx
@@ -1,6 +1,7 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components'
-import { CurrencyAmount, JSBI, Token, Trade } from '@pancakeswap/sdk'
+import { CurrencyAmount, Token, Trade } from '@pancakeswap/sdk'
+import { computeTradePriceBreakdown, warningSeverity } from 'utils/exchange'
 import {
   Button,
   Text,
@@ -19,11 +20,12 @@ import UnsupportedCurrencyFooter from 'components/UnsupportedCurrencyFooter'
 import Footer from 'components/Menu/Footer'
 import { useRouter } from 'next/router'
 import { useTranslation } from 'contexts/Localization'
-import { EXCHANGE_DOCS_URLS, BIG_INT_ZERO } from 'config/constants'
+import { EXCHANGE_DOCS_URLS } from 'config/constants'
+import { BIG_INT_ZERO } from 'config/constants/exchange'
 import { maxAmountSpend } from 'utils/maxAmountSpend'
-import { computeTradePriceBreakdown, warningSeverity } from 'utils/prices'
 import shouldShowSwapWarning from 'utils/shouldShowSwapWarning'
 import { useWeb3React } from '@web3-react/core'
+import { useSwapActionHandlers } from 'state/swap/useSwapActionHandlers'
 import useRefreshBlockNumberID from './hooks/useRefreshBlockNumber'
 import AddressInputPanel from './components/AddressInputPanel'
 import { GreyCard } from '../../components/Card'
@@ -47,7 +49,6 @@ import { Field } from '../../state/swap/actions'
 import {
   useDefaultsFromURLSearch,
   useDerivedSwapInfo,
-  useSwapActionHandlers,
   useSwapState,
   useSingleTokenSwapInfo,
 } from '../../state/swap/hooks'
